{{cookiecutter.project_slug}}/requirements/common.txt
Django
django-celery
django-environ
django-configurations
django-redis
djangorestframework-filters
djangorestframework
djangorestframework-jwt
psycopg2
coreapi
whitenoise
